handle directory downloads in DirectoryListingManager
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@964 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ windows/TreePropertySheet.cpp
@@ -102,7 +102,7 @@ void TreePropertySheet::fillTree() {
 	HTREEITEM initialItem = NULL;
 	for(int i = 0; i < pages; ++i) {
 		tab.GetItem(i, &item);
-		if(i == m_psh.nStartPage)
+		if(i == (int)m_psh.nStartPage)
 			initialItem = createTree(buf, TVI_ROOT, i);
 		else
 			createTree(buf, TVI_ROOT, i);
